FT5004 Project
In this project, we created a forum smart contract in which users can create many discussion threads. For token transfer matters within the forum, we created ForumToken to allow users to transfer tokens within the discussion. 

In each discussion thread, the discussion creator set the question title, content and reward (in Forum Token).